refactor: merge MessageThought to ThinkingBlock
